Moved MBC enum to gbstructs
